Capacity Planning Worksheet sample
In order to adequately calculate what velocity a team should plan to take into sprint, they utilize this worksheet to calculate their capacity for a given sprint.  This helps our teams to not over commit to work in a given sprint, because 100% of a few things is better than 80% of many nothings.